refactor(habitat): classify rules model by kind
