+++ tools/habitat-harness/src/providers/grit/docs-apply.ts
@@ -3,7 +3,7 @@ import {
   diagnosticCatalogEntryFromNativeRule,
   renderDiagnosticScanRootRefusal,
 } from "@internal/habitat-harness/service/model/check/index";
-import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
+import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/index";
 import { Effect } from "effect";
 import { docsLocalCheckoutPathsRewritePattern } from "./constants.js";
 import { GritProvider, type GritProviderRequirements } from "./resource.js";
