+++ tools/habitat-harness/src/providers/grit/failure.ts
@@ -3,7 +3,7 @@ import {
   renderDiagnosticProviderFailure,
 } from "@internal/habitat-harness/service/model/check/index";
 import type { RuleRunResult } from "@internal/habitat-harness/service/model/check/policy/rule-runtime/architecture";
-import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
+import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/index";
 
 export function infrastructureFailure(
   rule: RuleSourceFacts,
