+++ tools/habitat-harness/src/providers/grit/runner.ts
@@ -8,7 +8,7 @@ import {
   renderDiagnosticScanRootRefusal,
 } from "@internal/habitat-harness/service/model/check/index";
 import type { RuleRunResult } from "@internal/habitat-harness/service/model/check/policy/rule-runtime/architecture";
-import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
+import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/index";
 import { Effect } from "effect";
 import {
   gritDiagnosticOutcomesFromReport,
