+++ tools/habitat-harness/src/providers/grit/scan-roots/index.ts
@@ -5,7 +5,7 @@ import {
   renderDiagnosticScanRootRefusal,
 } from "@internal/habitat-harness/service/model/check/index";
 import { decideScanRootProtection } from "@internal/habitat-harness/service/model/check/index";
-import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
+import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/index";
 import { repoRoot, toRepoRelative } from "@internal/habitat-harness/resources/paths";
 import { protectedScanRootPrefixes } from "../constants.js";
