+++ tools/habitat-harness/src/service/model/check/policy/protected-zone-diagnostics.policy.ts
@@ -1,5 +1,5 @@
 import type { HabitatDiagnostic } from "@internal/habitat-harness/service/model/check/policy/structural/schema";
-import type { RuleFileLayerFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
+import type { RuleFileLayerFacts } from "@internal/habitat-harness/service/model/rules/index";
 import type {
   DeclarationReadiness,
   ProtectedMutationDecision,
