+++ tools/habitat-harness/src/service/model/check/policy/source/scan-roots.ts
@@ -1,6 +1,6 @@
 import path from "node:path";
 import { decideScanRootProtection } from "@internal/habitat-harness/service/model/check/index";
-import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
+import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/index";
 import { habitatCacheRepoPathPrefix } from "@internal/habitat-harness/resources/artifact-paths";
 import { repoRoot, toRepoRelative } from "@internal/habitat-harness/resources/paths";
