+++ tools/habitat-harness/src/service/model/rules/dto/registry.schema.ts
