+++ tools/habitat-harness/src/service/model/rules/index.ts
@@ -1,0 +1,6 @@
+export * from "./dto/registry.schema.js";
+export * from "./policy/artifact-paths.policy.js";
+export * from "./policy/facts.policy.js";
+export { ruleGraphFacts } from "./policy/graph.policy.js";
+export * from "./policy/path-coverage.policy.js";
+export * from "./repositories/registry.repository.js";
