--- tools/habitat-harness/src/service/model/rules/registry/index.ts
@@ -1,5 +1,0 @@
-export * from "./facts.js";
-export { ruleGraphFacts } from "./graph.js";
-export * from "./load.js";
-export * from "./path-coverage.js";
-export * from "./schema.js";
