+++ tools/habitat-harness/src/service/model/workspace/policy/workspace-targets.policy.ts
@@ -1,4 +1,4 @@
-import type { RuleGraphFacts } from "@internal/habitat-harness/service/model/rules/registry/schema";
+import type { RuleGraphFacts } from "@internal/habitat-harness/service/model/rules/dto/registry.schema";
 import { verifyAffectedTargetNames } from "@internal/habitat-harness/service/model/graph/policy/validation-routing.policy";
 import {
   type AggregateWorkspaceTargetDeclaration,
