+++ tools/habitat-harness/test/rules/registry/contract.test.ts
@@ -2,7 +2,7 @@ import { rules } from "@internal/habitat-harness/service/model/check/policy/rule
 import {
   parseRuleRegistryDocument,
   parseRuleRegistryText,
-} from "@internal/habitat-harness/service/model/rules/registry/index";
+} from "@internal/habitat-harness/service/model/rules/index";
 import { describe, expect, test } from "vitest";
 import { baseRule, expectInvalid, registryDocument } from "./helpers.js";
